fix: alert on x%

+++ src/coffee_maker_components/container.rs
@@ -18,7 +18,7 @@ pub trait Container {
         if max_capacity == 0 {
             false
         } else {
-            amount_percentage_alert >= (amount_left / max_capacity) as f32
+            amount_percentage_alert >= (amount_left as f32/ max_capacity as f32)
         }
     }
 
@@ -29,6 +29,7 @@ pub trait Container {
         max_capacity: u32,
         amount_percentage_alert: f32,
     ) -> bool {
+
         if self.amount_left_percentage_below_line(
             amount_left,
             max_capacity,
